Refactor
Making move to MVC pattern, away from God class of original design.

Add sceneGraph class to handle the contents of each scene. To render a scene the scene is simply passed to the render function.

Add controller class, allows for change in the control scheme without having to change the model, as per the design pattern requirements.

Add camera class to handle the data regarding camera position, direction and so on.