Refactor
Add view to mvc pattern,
this allows for multiple views to be created and displayed in the same window. So can create multiple cameras. This could be paired with multiple controllers will need further work but it looks good.

Add anti-aliasing

Seperated camera into own object, and is part of the view. So now the controller has access to the view and model and the view has access to the model.